Re #720 Fixed some GCC 4.6/7 warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@4975 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 495ee7ac3055ae5b413711c97c1dc439d96369fc
Former-commit-id: 0fffaa6e401ba055c7c59d2f05cae5cac81e1d5f

+++ src/libs/raceengine/raceengine.cpp
@@ -161,7 +161,7 @@ void RaceEngine::updateState(void)
 
 void RaceEngine::applyState(int state)
 {
-	::ReStateApply((void*)state);
+	::ReStateApply((void*)(long)state);
 }
 
 void RaceEngine::selectRaceman(GfRaceManager* pRaceMan, bool bKeepHumans)
