+++ src/libs/tgfclient/guiscrollbar.cpp
@@ -133,14 +133,14 @@ GfuiScrollBarCreate(void *scr, int x, int y, int length, int thickness, int butL
 								   "data/img/arrow-left-focused.png", "data/img/arrow-left-pushed.png",
 								   x, y, butLength, thickness,
 								   butMirror, false, 1,
-								   (void*)(object->id), gfuiScrollMinus,
+								   (void*)(long)(object->id), gfuiScrollMinus,
 								   NULL, (tfuiCallback)NULL, (tfuiCallback)NULL);
 			const tGfuiGrButton* pArrowBut = &(gfuiGetObject(scr, arrowButId)->u.grbutton);
 			GfuiGrButtonCreate(scr, "data/img/arrow-right.png", "data/img/arrow-right.png",
 							   "data/img/arrow-right-focused.png", "data/img/arrow-right-pushed.png",
 							   x + length - pArrowBut->width, y, butLength, thickness,
 							   butMirror, false, 1,
-							   (void*)(object->id), gfuiScrollPlus,
+							   (void*)(long)(object->id), gfuiScrollPlus,
 							   NULL, (tfuiCallback)NULL, (tfuiCallback)NULL);	    
 			break;
 		}
@@ -153,14 +153,14 @@ GfuiScrollBarCreate(void *scr, int x, int y, int length, int thickness, int butL
 								   "data/img/arrow-down-focused.png", "data/img/arrow-down-pushed.png",
 								   x, y, thickness, butLength,
 								   butMirror, false, 1,
-								   (void*)(object->id), gfuiScrollPlus,
+								   (void*)(long)(object->id), gfuiScrollPlus,
 								   NULL, (tfuiCallback)NULL, (tfuiCallback)NULL);	    
 			const tGfuiGrButton* pArrowBut = &(gfuiGetObject(scr, arrowButId)->u.grbutton);
 			GfuiGrButtonCreate(scr, "data/img/arrow-up.png", "data/img/arrow-up.png",
 							   "data/img/arrow-up-focused.png", "data/img/arrow-up-pushed.png",
 							   x, y + length - pArrowBut->height, thickness, butLength,
 							   butMirror, false, 1,
-							   (void*)(object->id), gfuiScrollMinus,
+							   (void*)(long)(object->id), gfuiScrollMinus,
 							   NULL, (tfuiCallback)NULL, (tfuiCallback)NULL);
 			break;
 		}
